refactor(web): extract TimeOptions component
